Add files via upload
I have written a python program to scrape the IMDB website and then load the desired data into an excel file.

Web Scraping is the process of programmatically extracting some data from a website. In Python, we can easily write programs to scrape websites by using the Beautiful Soup and Requests module. The requests module can be used to access the desired website and the Beautiful Soup module can be used to parse the HTML source code of the website. Beautiful soup makes it very easy to parse HTML content and then provides multiple methods which can be used to extract the data from any HTML tags. All the websites are written in HTML language hence in order to perform web scrapping, our program needs to read the HTML content. This is where Beautiful Soup comes into the picture to make it very easy to access the contents of HTML using very simple methods.